[REFACTOR]: Category 엔티티 클래스 패키지 변경
- board.domain -> category.domain
