+++ src/main/java/com/carrot/carrotmarketclonecoding/board/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.carrot.carrotmarketclonecoding.board.repository;
 
-import com.carrot.carrotmarketclonecoding.board.domain.Category;
+import com.carrot.carrotmarketclonecoding.category.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
